Docs: move docs-build-code to be importable
